More packaging updates from the master branch

+++ packaging/rpm/newscoop.spec
@@ -1,25 +1,24 @@
-# not for myself:
+# how-to:
 # http://fedoraproject.org/wiki/PackageMaintainers/CreatingPackageHowTo
 #
 %define manifest %{_builddir}/%{name}-%{version}-%{release}.manifest
 
 Summary:        The open content management system for professional journalists
 Name:           newscoop
-Version:        3.5.1
+Version:        4.1.0
 Release:        1
-License:        GPL
-Packager:       Robin Gareus <robin@gareus.org>
+License:        GPLv3
+Packager:       Daniel James <daniel@64studio.com>
 
-# TODO: This group does not seem right.
-# but the closest found in /usr/share/doc/rpm-*/GROUPS 
+# List found in /usr/share/doc/rpm-*/GROUPS 
 Group:          Applications/Publishing
 
 # TODO: use upstream tar ball for version > 3.5.2-rc2
 # we can not do that ATM because upstream has various issues.
 # see ../gen-debian-package.sh 
-#http://downloads.sourceforge.net/project/newscoop/$UPSTREAMDIST/newscoop-$UPSTREAMVERSION.tar.gz
+# http://downloads.sourceforge.net/project/newscoop/$UPSTREAMDIST/newscoop-$UPSTREAMVERSION.tar.gz
 Source0:        %{name}-%{version}.tar.gz
-URL:            http://www.sourcefabric.org/en/products/newscoop_overview/
+URL:            http://www.sourcefabric.org/en/newscoop/
 BuildRoot:      %{_tmppath}/%{name}-%{version}
 BuildArch:      noarch
 
@@ -31,13 +30,10 @@ Requires: php-mysql
 Requires: curl
 Requires: mysql
 Requires: ImageMagick
-# TODO: find MTA package
-#Requires: mail-server
-#Requires: sendmail
 
-# These are actually recommends only:
-#Requires: mysql-server
-#Requires:: php-apc | php5-xcache
+#These are recommends, rpm does not fully support the Suggests tag though:
+#Suggests: mysql-server
+#Suggests: postfix
 
 %description
 Newscoop is the open content management system for professional journalists.
@@ -56,20 +52,18 @@ rm -rf %{buildroot}
 mkdir -p %{buildroot}/var/lib/
 cp -a newscoop %{buildroot}/var/lib/
 
-# TODO: create config-files - debian/ folder will not be present in tar-ball
-mkdir -p %{buildroot}/etc/newscoop/3.5/
-cp debian/etc/newscoop.ini %{buildroot}/etc/newscoop/3.5/
-cp debian/etc/apache.conf %{buildroot}/etc/newscoop/3.5/
-cp debian/etc/apache.vhost.tpl %{buildroot}/etc/newscoop/3.5/
-cp debian/etc/newscoop.cron.tpl %{buildroot}/etc/newscoop/3.5/
+# Copy config file
+mkdir -p %{buildroot}/etc/newscoop/4.1/
+cp rpm/newscoop.ini %{buildroot}/etc/newscoop/4.1/
+cp rpm/apache.conf %{buildroot}/etc/newscoop/4.1/
 
 cd $RPM_BUILD_ROOT
 rm -f %{manifest}
-find ./var/ -type d \
-        | sed '1,2d;s,^\.,\%attr(-\,apache\,apache) \%dir ,' >> %{manifest}
-find ./var/ -type f \
+find ./var/lib/ -type d \
+        | sed '1d;s,^\.,\%attr(-\,apache\,apache) \%dir ,' >> %{manifest}
+find ./var/lib/ -type f \
         | sed 's,^\.,\%attr(-\,apache\,apache) ,' >> %{manifest}
-find ./var/ -type l \
+find ./var/lib/ -type l \
         | sed 's,^\.,\%attr(-\,apache\,apache) ,' >> %{manifest}
 
 %clean
@@ -78,15 +72,15 @@ rm -rf %{buildroot}
 
 %files -f %{manifest}
 %defattr(-,root,root)
-%doc ChangeLog  COPYING  CREDITS README  UPGRADE
-%config /etc/newscoop/3.5/apache.conf
-%config /etc/newscoop/3.5/newscoop.ini
-%config /etc/newscoop/3.5/apache.vhost.tpl
-%config /etc/newscoop/3.5/newscoop.cron.tpl
+%doc ChangeLog CREDITS README  UPGRADE
+%config /etc/newscoop/4.1/apache.conf
+%config /etc/newscoop/4.1/newscoop.ini
+#%config /etc/newscoop/4.1/apache.vhost.tpl
+#%config /etc/newscoop/4.1/newscoop.cron.tpl
 
 %post
 # symlink config files
-configdir="/etc/newscoop/3.5"
+configdir="/etc/newscoop/4.1"
 includefile="${configdir}/apache.conf"
 phpinifile="${configdir}/newscoop.ini"
 webserver="httpd"
@@ -96,6 +90,7 @@ dohtaccess="/newscoop"
 if [ ! -d /etc/$webserver/conf.d/ ]; then
 		install -d -m755 /etc/$webserver/conf.d/
 fi
+
 if [ ! -e /etc/$webserver/conf.d/newscoop.conf ]; then
 	ln -s ${includefile} /etc/$webserver/conf.d/newscoop.conf
 fi
@@ -103,20 +98,24 @@ fi
 if [ ! -d /etc/$php/conf.d/ ]; then
 	install -d -m755 /etc/php.d/
 fi
+
 if [ ! -e /etc/php.d/newscoop.ini ]; then
 	ln -s ${phpinifile} /etc/php.d/newscoop.ini
 fi
 
 # .htaccess file
-echo -ne "/RewriteBase/d\nwq\n\n" \
-| ed /var/lib/newscoop/.htaccess &>/dev/null || true
+#echo -ne "/RewriteBase/d\nwq\n\n" \
+#| ed /var/lib/newscoop/.htaccess &>/dev/null || true
+#
+#if [ -n "${dohtaccess}" ]; then
+#	echo -ne "/RewriteEngine/\n+1i\n    RewriteBase ${dohtaccess}\n.\nwq\n" \
+#	| ed /var/lib/newscoop/.htaccess &>/dev/null || true
+#fi
 
-if [ -n "${dohtaccess}" ]; then
-	echo -ne "/RewriteEngine/\n+1i\n    RewriteBase ${dohtaccess}\n.\nwq\n" \
-	| ed /var/lib/newscoop/.htaccess &>/dev/null || true
-fi
+# Fix SELinux
+chcon -R -t httpd_cache_t /var/lib/newscoop
 
-# XXX: restart apache - check if this is the recommended way
+# restart Apache
 /etc/init.d/httpd restart
 
 ## CRON JOB
@@ -133,27 +132,55 @@ fi
 
 %postun
 webserver="httpd"
-if [ -L /etc/$webserver/conf.d/newscoop.conf ]; then
-	rm -f /etc/$webserver/conf.d/newscoop.conf || true
-fi
-		
-if [ -L /etc/php.d/newscoop.ini ]; then
-	rm -f /etc/php.d/newscoop.ini || true
-fi
 
-if [ -L /etc/cron.d/newscoop ]; then
-	rm -f /etc/cron.d/newscoop || true
-fi
-# delete generated templates and user-installed plugins
-rm -rf /var/lib/newscoop || true
-rm -f /etc/newscoop/3.5/newscoop.cron || true
-rmdir /etc/newscoop/3.5 || true
-rmdir /etc/newscoop/ || true
+# delete Newscoop files, but only on uninstallation
+if [ "$1" = "0" ]; then
+
+ if [ -L /etc/$webserver/conf.d/newscoop.conf ]; then
+	rm -f /etc/$webserver/conf.d/newscoop.conf || true
+ fi
 		
-# XXX: restart apache - check if this is the recommended way
+ if [ -L /etc/php.d/newscoop.ini ]; then
+	rm -f /etc/php.d/newscoop.ini || true
+ fi
+
+ if [ -L /etc/cron.d/newscoop ]; then
+	rm -f /etc/cron.d/newscoop || true
+ fi
+
+ rm -rf /var/lib/newscoop/ || true
+ rm -rf /etc/newscoop/ || true
+
+fi
+		
+# restart Apache
 /etc/init.d/httpd restart
 
 
 %changelog
+* Mon Jan 28 2013 Daniel James <daniel@64studio.com>
+- Update for Newscoop 4.1.0
+
+* Mon Jan 7 2013 Daniel James <daniel@64studio.com>
+- Update for Newscoop 4.0.4
+
+* Wed Dec 13 2012 Daniel James <daniel@64studio.com>
+- Update for Newscoop 4.0.3
+
+* Wed Jul 18 2012 Daniel James <daniel@64studio.com>
+- Update for Newscoop 4.0.2
+
+* Wed Jun 20 2012 Daniel James <daniel@64studio.com>
+- Test for upgrade or uninstall before deleting files
+
+* Tue Jun 19 2012 Daniel James <daniel@64studio.com>
+- Update for Newscoop 4.0.1
+
+* Tue May 8 2012 Daniel James <daniel@64studio.com>
+- Put DocumentRoot in /var/lib as /usr may be read-only
+
+* Mon Apr 30 2012 Daniel James <daniel@64studio.com>
+- Update for Newscoop 4.0.0
+
 * Wed Jan 26 2011 Robin Gareus <robin@gareus.org>
 - Initial Version
